[ENSCORESW-595]. Creating a new export command.
This has been adapted from another developer's code. It is being
editted to allow for more of a user interactive mode. We include
checking if submitted refs are part of the current branch. We also
make sure that the directories are resolved fully and store our
config in the git repository's config. The original pushed everything
into a file.