[ENSCORESW-595]. Do parent sniffing to select the correct commit.
Doing the rev-list from our last commit will give all commits that happened
since that commit. This is not limited to a single branch and does not include
the parent commit. Our solution is to use rev-list and look for the most
leftly reported commit. Experience shows this will be the branch the merge was
pushed into (master). We can then skip any commit that is not a child
of our last pushed hash.

Subsequently post CVS commit our last pushed hash becomes the commit we have just
sent in.

Also we have lots more documentation about the commit process and what is allowed