[ENSCORESW-595]. Switched to using --first-parent for history
Seems that using --first-parent rather than --parents was the answer. This
option lets me see the first parent of a merge commit meaning I only get
to see a single track of history. That also means we do not need to check
the parent is right because it must be. Simplification and happiness